+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased changes
+- Support for `GetBakerEarliestWinTime`. Requires node version at least 6.1.
 - Add a `commission_rates` field to `CurrentPaydayBakerPoolStatus` which yields the commission rates 
   of the baker for the reward period (requires a node with version at least 6.1).
 - Add support for `GetBakersRewardPeriod` endpoint. Requires a node with version at least 6.1.