+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased changes
 - Add a `commission_rates` field to `CurrentPaydayBakerPoolStatus` which yields the commission rates 
   of the baker for the reward period (requires a node with version at least 6.1).
-- Add support for `GetBlockCertificates`. Requires a node with version at least 6.1.
+- Add support for `GetBlockCertificates`. Requires a node version at least 6.1.
 
 ## 3.0.0