+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased changes
 - Add a `commission_rates` field to `CurrentPaydayBakerPoolStatus` which yields the commission rates 
-  of the baker for the reward period (requires a node with version at least 6.1).
+  of the baker for the reward period. Requires a node version at least 6.1.
+- Add support for `GetWinningBakersEpoch`. Requires a node version at least 6.1.
+- Add Support for `GetFirstBlockEpoch`. Requires a node version at least 6.1.
 - Add support for `GetBakersRewardPeriod` endpoint. Requires a node with version at least 6.1.
 - Add Support for `GetBakerEarliestWinTime` endpoint. Requires a node with version at least 6.1.